Question 2.1

+++ src/Controller/CharactersController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Characters;
 use App\Form\CharactersType;
 use App\Repository\CharactersRepository;
+use App\Upload\FileCharactersUpload;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -28,13 +29,15 @@ class CharactersController extends AbstractController
     /**
      * @Route("/new", name="characters_new", methods={"GET","POST"})
      */
-    public function new(Request $request): Response
+    public function new(Request $request,FileCharactersUpload $fileCharactersUpload): Response
     {
         $character = new Characters();
         $form = $this->createForm(CharactersType::class, $character);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $fileCharactersUpload->upload($character);
+
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($character);
             $entityManager->flush();
@@ -61,12 +64,14 @@ class CharactersController extends AbstractController
     /**
      * @Route("/{id}/edit", name="characters_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, Characters $character): Response
+    public function edit(Request $request, Characters $character,FileCharactersUpload $fileCharactersUpload): Response
     {
         $form = $this->createForm(CharactersType::class, $character);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $fileCharactersUpload->upload($character);
+
             $this->getDoctrine()->getManager()->flush();
 
             return $this->redirectToRoute('characters_index', [
@@ -83,12 +88,15 @@ class CharactersController extends AbstractController
     /**
      * @Route("/{id}", name="characters_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, Characters $character): Response
+    public function delete(Request $request, Characters $character,FileCharactersUpload $fileCharactersUpload): Response
     {
         if ($this->isCsrfTokenValid('delete'.$character->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($character);
-            $entityManager->flush();
+            if($fileCharactersUpload->removeFile($character) === true) {
+
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($character);
+                $entityManager->flush();
+            }
         }
 
         return $this->redirectToRoute('characters_index');
